Version 1.1 of the Hosted Payments Library :confetti_ball:
Version 1.1 is a significant rewrite of the Hosted Payments library,
designed to make the library easier to use with Composer and modern
PHP frameworks.

The updated version of the library now respects PSR-1, PSR-2,
and uses PSR-4 autoloading.  For more information, see:
http://www.php-fig.org/psr/

Extensive updates to in-code documentation should make code-completion
in IDEs like PHPStorm way more useful as well.